Flask app: add ability to keep maps on the server for a gallery. Performance improvements gained from removing redundant information client side.

+++ manifest_data/__init__.py
@@ -9,8 +9,8 @@
 
 ## pygmaps carries the Apache 2.0 license
 
-import re
 import os
+import re
 import random
 from uuid import uuid4
 from flask import Flask, render_template, request
@@ -20,12 +20,6 @@ import pygmaps
 app = Flask(__name__)
 app.config['MAX_CONTENT_LENGTH'] = 6 * 1024 * 1024
 
-#   Regular expression for splitting filename
-spl = re.compile(
-    r"(?P<ts>\d+)T"
-    r"(?P<orig>\d{3}\.\d{3}\.\d{3}\.\d{3})\.(?P<orig_port>\d+)\-"
-    r"(?P<dest>\d{3}\.\d{3}\.\d{3}\.\d{3})\.(?P<dest_port>\d+)"
-)
 
 @app.route('/')
 def root():
@@ -37,39 +31,20 @@ def google_map_form():
     if request.method == 'POST':
         file = request.files.get('file')
         if file:
-            # if file.mimetype == 'txt/plain':
+            keep = request.post.get('keep')
             data = file.read().split()
-            output = geolocate_ips(data)
+            output = geolocate_ips(data, keep)
             return render_template('map_output.html', output=output)
     else:
         return render_template('file_upload.html')
 
-class LocData:
-    """
-    data wrapper for geolocation data
-    """
 
-    def __init__(self, **kwargs):
-        self.orig = self.str_format(kwargs['orig'])
-        self.dest = self.str_format(kwargs['dest'])
+regex_validator = re.compile('^{r}\.{r}\.{r}\.{r},'
+                             '{r}\.{r}\.{r}\.{r}$'.format(r='\d{1,3}')
+)
 
-    def str_format(self, s):
-        """
-        strips leading 0 from IP address blocks
-        """
-        a = s.split('.')
-        return '.'.join([z.lstrip('0') for z in a])
 
-def split_name(f):
-    """
-    splits filename according to regex defined above
-    """
-    split = spl.search(f)
-    if split:
-        return split.groupdict()
-    return None
-
-def geolocate_ips(files):
+def geolocate_ips(files, keep):
     """
     create google map of geolocated IP addresses with paths
     connecting points
@@ -82,43 +57,52 @@ def geolocate_ips(files):
     try:
         locobj.open(loc_path)
     except:
-        raise Exception('Could not find IP2Location binary... '
-                        'Make sure it is in {0}/data'.format(BASE_DIR))
+        raise Exception('Could not find IP2Location binary... ')
 
     #   Creates map with a center roughly over the US
     mymap = pygmaps.maps(37.428, -122.145, 4)
 
     s = []
     for f in files:
-        dct = split_name(f)
-        if dct:
-            d = LocData(**dct)
-            orig_latlong = geolocate(locobj, d.orig)
-            dest_latlong = geolocate(locobj, d.dest)
-            out = [orig_latlong, dest_latlong]
-            if out not in s:
-                s.append(out)
+        if not regex_validator.search(f):
+            continue
+        orig, dest = f.split(',')
+        orig_latlong = geolocate(locobj, orig)
+        dest_latlong = geolocate(locobj, dest)
+        out = [orig_latlong, dest_latlong]
+        s.append(out)
     for path in s:
         if (0.0, 0.0) not in path:
             #	sets random hex color for gmaps path
             r = lambda: random.randint(0, 255)
             h = '#%02X%02X%02X' % (r(), r(), r())
             mymap.addpath(path, h)
-    output = os.path.join(os.path.dirname(__file__), 'tmp', uuid4())
-    mymap.draw(output)
+    addendum = """
+                <html>
+                    <body>
+                        <p style="text-align: center">The IP geolocation data is
+                        provided by <a href="http://www.ip2location.com">
+                        http://www.ip2location.com</a></p>
+                    </body>
+                </html>"""
+    if keep:
+        output = os.path.join(os.path.dirname(__file__), 'static/gallery',
+                              str(uuid4()) + '.html')
+        mymap.draw(output)
+        #	for attribution of IP data
+        with open(output, 'ab') as fi:
+            fi.write(addendum)
+        with open(output, 'rb') as fi:
+            html = fi.read()
+    else:
+        output = os.path.join(os.path.dirname(__file__), 'tmp', uuid4())
+        mymap.draw(output)
 
-    #	for attribution of IP data
-    with open(output, 'rb') as fi:
-        html = fi.read()
-        html += """
-			<html>
-				<body>
-					<p style="text-align: center">The IP geolocation data is
-					provided by <a href="http://www.ip2location.com">
-					http://www.ip2location.com</a></p>
-				</body>
-			</html>"""
-    os.unlink(output)
+        #	for attribution of IP data
+        with open(output, 'rb') as fi:
+            html = fi.read()
+            html += addendum
+        os.unlink(output)
     return html
 
 
@@ -138,4 +122,4 @@ def geolocate(locobj, ip):
     return (req.latitude, req.longitude)
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    app.run()
